backend/requirements.txt
django==2.2
djangorestframework==3.9.2
django-cors-headers==2.5.2
django-filter==2.1.0
Markdown==3.1
pytz==2019.1
sqlparse==0.3.0
